refactor(execution): move prompt builders into RunInteractionService
Phase 4e of the CQRS refactor (#234, third of three commits). Finishes
the Phase 4d agent-lifecycle boundary that was left partial.

The setup + do-work prompts are an implementation detail of HOW to
talk to the agent — they belong with the agent lifecycle owner, not
on the orchestrator. Phase 4d consciously deferred the signature
change to keep that PR bounded. Phase 4e completes it:

  - Move `buildSetupPrompt`, `buildDoWorkPrompt`, `buildPrompt`
    verbatim from ExecutionService into RunInteractionService
  - Update `RunSessionInput`: drop the pre-built `setupPrompt` /
    `doWorkPrompt` string fields, add raw materials `node`,
    `workItem`, `issueBody`, `projectContext`
  - `runSetupPhase` and `runDoWorkPhase` now call their own
    `buildSetupPrompt` / `buildDoWorkPrompt` helpers internally
  - `ExecutionService.executeRun` drops the two pre-build calls
    and passes the raw materials in the input envelope
  - `ExecutionService.launch` + the run-record creation path rewire
    the legacy-compat `this.buildPrompt(workItem, node)` call sites
    to `this.runInteractionService.buildPrompt(workItem, node)`
  - Drop the dead `workItemSlug` import from ExecutionService

ExecutionService line count drops 935 → 791 (−144 from this commit;
−358 from this Phase 4e PR; −1092 from the start of the Phase 4 epic
when ExecutionService was ~1883 lines).

- npm run check ✓ (tsc --noEmit)
- npx vitest run --no-file-parallelism ✓ (561 / 561)
- forwardRef count unchanged: 8
